+++ changelog.md
@@ -1,3 +1,1 @@
-- added compat with camera mod
-- added compat with refurbished mod
-- fixed forge energy not updating client
+- more compat hacks